# server/spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Prompt Mint API
  version: "1.0.0"
  description: |
    Prompt Mint marketplace API — Express backend serving prompt listing,
    user management, governance, appeals, licensing, webhooks, versioning,
    notifications, analytics, and API key management.
  contact:
    name: Prompt Mint Team
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.promptmint.io
    description: Production server
paths:
...
  /api/seo/controls:
    get:
      operationId: getSEOControls
      summary: Get SEO control headers
      tags: [SEO]
      responses:
        "200":
          description: SEO control settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SEOControls"
    post:
      operationId: updateSEOControls
      summary: Update SEO control headers
      tags: [SEO]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SEOControls"
      responses:
        "200":
          description: SEO controls updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SEOControls"

  # ─── Improve Proxy ─────────────────────────────────────────────────────────
...
  /api/user/export/challenge:
    post:
      operationId: generateExportChallenge
      summary: Generate a Stellar challenge for data export
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
      responses:
        "200":
          description: Challenge generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
  /api/user/export:
    post:
      operationId: requestExport
      summary: Request a signed data export
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                signature:
                  type: string
                challenge:
                  type: string
      responses:
        "200":
          description: Export initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  exportId:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
  /api/user/export/download/{exportId}:
    get:
      operationId: downloadExport
      summary: Download a completed data export
      tags: [Users]
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export file
          content:
            application/json:
              schema:
                type: object
  /api/user/delete/challenge:
    post:
      operationId: generateDeletionChallenge
      summary: Generate a Stellar challenge for account deletion (#91)
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
      responses:
        "200":
          description: Challenge generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  challenge:
                    type: string
                  expiresAt:
                    type: number
                  nonce:
                    type: string
  /api/user/delete:
    post:
      operationId: requestAccountDeletion
      summary: Delete the off-chain profile data for a wallet (#91)
      description: >-
        Signature-gated deletion of the caller's User profile, notifications,
        and webhook subscriptions. Marketplace purchase/order records, votes,
        and moderation reports are intentionally retained to preserve
        on-chain access authority and audit integrity -- see
        docs/legal/data-retention-policy.md.
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                signature:
                  type: string
                token:
                  type: string
      responses:
        "200":
          description: Off-chain profile data deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: object
                    properties:
                      profile:
                        type: boolean
                      webhookSubscriptions:
                        type: number
                      notifications:
                        type: number
                  retained:
                    type: object
                    properties:
                      collections:
                        type: array
                        items:
                          type: string
                      reason:
                        type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ─── Versions ──────────────────────────────────────────────────────────────
...
  /api/notifications/export:
    get:
      operationId: exportNotifications
      summary: Export the full notification history for a wallet
      description: >
        Returns every notification for the wallet (read and unread, newest
        first) as a downloadable attachment. Use `format=csv` for CSV or the
        default `format=json` for JSON.
      tags: [Notifications]
      parameters:
        - name: walletAddress
          in: query
          required: true
          description: Wallet address that owns the notification history.
          schema:
            type: string
        - name: format
          in: query
          required: false
          description: Export file format. Defaults to json.
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: Notification history attachment (CSV or JSON)
          content:
            application/json:
              schema:
                type: object
                properties:
                  walletAddress:
                    type: string
                  exportedAt:
                    type: string
                    format: date-time
                  count:
                    type: integer
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
            text/csv:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: Admin API token from ADMIN_API_TOKEN env var
    apiKeyQuery:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key for programmatic access
  responses:
    BadRequest:
      description: Bad request — missing or invalid fields
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
...
  schemas:
...
    SEOControls:
      type: object
      properties:
        xRobotsTag:
          type: string
          example: noindex

    # ── Prompt ────────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        data:
          type: object
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ── Prompt Order ──────────────────────────────────────────────────────────
...